+++ maths-club-app/package.json
@@ -39,7 +39,7 @@
     "@angular/language-service": "~9.1.3",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
+    "@types/node": "^14.0.27",
     "codelyzer": "^5.1.2",
     "eslint": "^7.2.0",
     "eslint-config-prettier": "^6.11.0",
