+++ maths-club-app/src/app/services/notification.service.ts
@@ -1,0 +1,56 @@
+import { Injectable } from "@angular/core";
+import {
+  Plugins,
+  PushNotification,
+  PushNotificationToken,
+  PushNotificationActionPerformed,
+} from "@capacitor/core";
+
+const { PushNotifications, Modals } = Plugins;
+
+@Injectable({
+  providedIn: "root",
+})
+export class NotificationService {
+  constructor() {}
+
+  init() {
+    PushNotifications.requestPermission().then((result) => {
+      if (result.granted) {
+        PushNotifications.register();
+      }
+    });
+
+    PushNotifications.addListener(
+      "registration",
+      (token: PushNotificationToken) => {
+        alert("Push registration success, token:" + token.value);
+        console.log("Token Value", token.value);
+      }
+    );
+
+    PushNotifications.addListener("registrationError", (error: any) => {
+      alert("Error on registration: " + JSON.stringify(error));
+    });
+
+    PushNotifications.addListener(
+      "pushNotificationReceived",
+      (notification: PushNotification) => {
+        // alert("Push received: " + JSON.stringify(notification));
+        console.log("Push Received: ", notification);
+
+        let alertRet = Modals.alert({
+          title: notification.title,
+          message: notification.body,
+        });
+      }
+    );
+
+    PushNotifications.addListener(
+      "pushNotificationActionPerformed",
+      (notification: PushNotificationActionPerformed) => {
+        alert("Push action performed: " + JSON.stringify(notification));
+      }
+    );
+  }
+}
