+++ maths-club-app/src/environments/environment.prod.ts
@@ -1,4 +1,6 @@
+import PACKAGE_JSON from "../../package.json";
+
 export const environment = {
-  production: true,
-  APP_VERSION: require("../../package.json").version,
+  production: false,
+  APP_VERSION: PACKAGE_JSON.version,
 };
